Fix hyperparameters and added further results.

+++ training/hyper/hyper_qm7.py
@@ -214,7 +214,7 @@ hyper = {
             "config": {
                 "name": "PAiNN",
                 "inputs": [
-                    {"shape": [None], "name": "node_number", "dtype": "float32", "ragged": True},
+                    {"shape": [None], "name": "node_number", "dtype": "int64", "ragged": True},
                     {"shape": [None, 3], "name": "node_coordinates", "dtype": "float32", "ragged": True},
                     {"shape": [None, 2], "name": "range_indices", "dtype": "int64", "ragged": True}
                 ],
